koreksi

+++ app/Http/Livewire/Bacameter/Buattarget.php
@@ -17,20 +17,6 @@ class Buattarget extends Component
         $this->tahun = $this->tahun ?: date('Y');
     }
 
-    public function setProses($proses = null)
-    {
-        $this->validate([
-            "bulan" => "required",
-            "tahun" => "required",
-        ]);
-
-        if (BacaMeter::where('periode', $this->tahun . "-" . $this->bulan . "-01")->count() > 0) {
-            session()->flash('danger', 'Data bacaan periode ' . $this->tahun . '-' . $this->bulan . ' ini sudah ada');
-        }
-
-        $this->proses = $proses;
-    }
-
     public function submit()
     {
         ini_set('max_execution_time', 0);
@@ -41,39 +27,47 @@ class Buattarget extends Component
             "bulan" => "required",
             "tahun" => "required",
         ]);
+        $posting = true;
 
-        DB::transaction(function ($q) {
-            BacaMeter::whereNull('tanggal_baca')->where('periode', $this->tahun.'-'.$this->bulan.'-01')->forceDelete();
-            $pelanggan = Pelanggan::with('bacaMeterTerakhir')->with('jalan.rayonDetail.rayon.ruteBaca')->whereNotIn('id', BacaMeter::where('periode', $this->tahun.'-'.$this->bulan.'-01')->get()->pluck('pelanggan_id')->all())->whereIn('status', [1, 3])->get();
-            $data = [];
-            foreach ($pelanggan as $key => $row) {
-                array_push($data, [
-                    'periode' => $this->tahun . '-' . $this->bulan . '-01',
-                    'stand_lalu' => $row->bacaMeterTerakhir ? $row->bacaMeterTerakhir->stand_ini : 0,
-                    'latitude' => $row->latitude,
-                    'longitude' => $row->longitude,
-                    'pelanggan_id' => $row->id,
-                    'pembaca_id' => $row->jalan->rayonDetail->rayon->ruteBaca->pembaca_id,
-                    'jalan_id' => $row->jalan_id,
-                    'rayon_id' => $row->jalan->rayonDetail->rayon_id,
-                    'pengguna_id' => auth()->id(),
-                    'created_at' => now(),
-                    'updated_at' => now(),
+        if (BacaMeter::where('periode', $this->tahun . "-" . $this->bulan . "-01")->count() > 0) {
+            session()->flash('danger', 'Data bacaan periode ' . $this->tahun . '-' . $this->bulan . ' ini sudah ada');
+            $posting = false;
+        }
+
+        if ($posting == true) {
+            DB::transaction(function ($q) {
+                BacaMeter::whereNull('tanggal_baca')->where('periode', $this->tahun . '-' . $this->bulan . '-01')->forceDelete();
+                $pelanggan = Pelanggan::with('bacaMeterTerakhir')->with('jalan.rayonDetail.rayon.ruteBaca')->whereNotIn('id', BacaMeter::where('periode', $this->tahun . '-' . $this->bulan . '-01')->get()->pluck('pelanggan_id')->all())->whereIn('status', [1, 3])->get();
+                $data = [];
+                foreach ($pelanggan as $key => $row) {
+                    array_push($data, [
+                        'periode' => $this->tahun . '-' . $this->bulan . '-01',
+                        'stand_lalu' => $row->bacaMeterTerakhir ? $row->bacaMeterTerakhir->stand_ini : 0,
+                        'latitude' => $row->latitude,
+                        'longitude' => $row->longitude,
+                        'pelanggan_id' => $row->id,
+                        'pembaca_id' => $row->jalan->rayonDetail->rayon->ruteBaca->pembaca_id,
+                        'jalan_id' => $row->jalan_id,
+                        'rayon_id' => $row->jalan->rayonDetail->rayon_id,
+                        'pengguna_id' => auth()->id(),
+                        'created_at' => now(),
+                        'updated_at' => now(),
+                    ]);
+                }
+
+                $insert = collect($data)->chunk(2000);
+                foreach ($insert as $row) {
+                    BacaMeter::insert($row->toArray());
+                }
+
+                BacaMeter::where('pelanggan_id', $pelanggan->where('status', 3)->pluck('id')->all())->where('periode', $this->tahun . '-' . $this->bulan . '-01')->update([
+                    'stand_ini' => DB::raw('stand_lalu'),
+                    'status_baca' => 'PUTUS SEMENTAR PERMINTAAN SENDIRI',
+                    'tanggal_baca' => now(),
                 ]);
-            }
-
-            $insert = collect($data)->chunk(2000);
-            foreach ($insert as $row) {
-                BacaMeter::insert($row->toArray());
-            }
-
-            BacaMeter::where('pelanggan_id', $pelanggan->where('status', 3)->pluck('id')->all())->where('periode', $this->tahun . '-' . $this->bulan . '-01')->update([
-                'stand_ini' => DB::raw('stand_lalu'),
-                'status_baca' => 'PUTUS SEMENTAR PERMINTAAN SENDIRI',
-                'tanggal_baca' => now(),
-            ]);
-            session()->flash('success', 'Data bacaan periode ' . $this->tahun . '-' . $this->bulan . ' berhasil dibuat');
-        });
+                session()->flash('success', 'Data bacaan periode ' . $this->tahun . '-' . $this->bulan . ' berhasil dibuat');
+            });
+        }
         $this->reset('proses');
     }
